Refine the GitHub Actions Cache hack
Previously, the commit SHA sum was used for the cache key.
The result of this was that if the fuzzing workflow ran on the same
commit as the previous run (i.e., there were no new commits on master in
the past day), that the cache wouldn't be updated.

Now, use the GitHub Actions Run ID instead.  The Run ID is supposed to
be a unique number for each run within a repository, that doesn't change
if you re-run the workflow.  This should cause the cache to be updated
more frequently.

What would be more clear would be to use an explicit cache, perhaps an
S3 bucket -- see #76.