Add methods for a barebones IR visitor protocol.
Oftentimes, a visitor protocol _also_ abstracts over traversal order.
However, it's hard to choose a generic traversal order that's suitable
for all your use cases, so what you usually see is either _multiple_
visitor protocols for different traversal orders.

Instead of dealing with all that, the idea here is that each visitor
implementation will be responsible in its `.Visit(const Foo &foo)`
methods to traverse the subcomponents of `foo` in the desired order.