Fix several bugs in reproc.
In particular:

- Several places misused assertions for runtime error checking, via the
  `ASSERT_UNUSED` macro.  These have been replaced with more correct
  error handling.

- Nowhere in reproc were interrupted system calls explicitly handled.
  Many of the places that were misusing assertions in fact needed a loop
  to repeat the relevant code if it was interrupted.

Note: checking in fixes directly on top of the vendored code like this
is not great, particularly with the squash-on-merge workflow we are
using for this project.  These combined could make it difficult to see
what modifications to third-party code have been made!

I am also currently trying to get these fixes accepted upstream.  If and
when they are accepted, we can update this code to use a newer version.
In the meantime, we need these fixes for the new fuzzing code to work
correctly.